# Speed up clippy