--- a/scripts/src/test.sh
+++ b/scripts/src/test.sh
@@ -90,7 +90,7 @@ runtime_upgrade_test() {
 
   # Run node
   RUST_LOG="pallet_gear=debug,runtime::gear::hooks=debug" $ROOT_DIR/target/release/gear-node \
-  --dev --tmp --unsafe-ws-external --unsafe-rpc-external --rpc-methods Unsafe --rpc-cors all & sleep 2
+  --dev --tmp --unsafe-ws-external --unsafe-rpc-external --rpc-methods Unsafe --rpc-cors all & sleep 10
 
   # Change dir to the js script dir
   cd "$TEST_SCRIPT_PATH"