### Changes
Compiler takes a lot of time for rebuilding wasms, but this is excess for `clippy` and `check` so this got removed.
For more details, see https://github.com/gear-tech/gear/issues/467#issuecomment-1104002817

* Wasm build skipped for clippy and check
* Removed legacy from makefile
* Removed unused runtime pallet implementation
* Bumped sleep time from 2 to 10 in runtime upgrade test to avoid hardware slows which lead to infinity test

### Note
This also affects CI, due to env variable declaring directly in scripts

### Performance changes _(before / after ≈ performance_increasement)_
Examples
**39,8s / 34,9s ≈ +12,4%**

Workspace
**3m 55,5s / 3m 26,1s ≈ +12.5%**

Rebuild workspace (minor changes in gear pallet)
**4,8s / 3,6s ≈ +15%**